qemu-timer.c: add comments to note meaning of "REALTIME" and "HOST" clocks
Add some simple comments to note that QEMU_CLOCK_REALTIME really refers to
a monotonically increasing clock, and that QEMU_CLOCK_HOST refers to the
host's realtime clock (either gettimeofday() or clock_gettime(CLOCK_REALTIME).
And for those not in the know, QEMU_CLOCK_VIRTUAL refers to the VM's clock.

Signed-off-by: Peter Portante <peter.portante@redhat.com>